% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gpr_posterior}
\alias{gpr_posterior}
\title{GPR posterior}
\usage{
gpr_posterior(x, y, x.targets, noise, kernelfunc, derivatives = FALSE)
}
\arguments{
\item{x}{obs timepoints}

\item{y}{obs values}

\item{x.targets}{target timepoints}

\item{noise}{noise std, a single value or a vector}

\item{kernelfunc}{a kernel function of (x1,x2) (returns a matrix)}

\item{derivatives}{compute also derivatives}
}
\value{
a \code{gpsimple}-object with fields
\item{x}{timepoints}
\item{mean}{GP mean}
\item{cov}{covariance matrix}
\item{noisestd}{vector of noise std's}
\item{mll}{marginal log likelihood}
\item{x.obs}{original observation times}
\item{y.obs}{original observation values}
}
\description{
a simple GPR posterior distribution, no parameter learning
}
\keyword{internal}

