% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gpsimple}
\alias{gpsimple}
\title{gpsimple object, a simple Gaussian Process container with fields}
\usage{
gpsimple(x, mean, cov, noisestd = NA, mll = NA, x.obs = NA, y.obs = NA)
}
\arguments{
\item{x}{inputs}

\item{mean}{mean vector}

\item{cov}{covariance matrix}

\item{noisestd}{noise std vector}

\item{mll}{marginal log likelihood of data}

\item{x.obs}{observed input times}

\item{y.obs}{observed input values}
}
\value{
a \code{gpsimple} object with same fields as parameters
}
\description{
gpsimple object, a simple Gaussian Process container with fields
}
\keyword{internal}

