% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.gpsimple}
\alias{plot.gpsimple}
\title{Plot a simple gaussian process}
\usage{
\method{plot}{gpsimple}(x, y = NULL, plotdata = TRUE, plotmean = TRUE,
  plotcov = TRUE, plotnoise = FALSE, samples = 0, sigma = 2,
  title = NULL, legend = FALSE, plotgradient = TRUE, ...)
}
\arguments{
\item{x}{the gp-object}

\item{y}{placeholder variable}

\item{plotdata}{plot the data (default)}

\item{plotmean}{plot the GP mean (default)}

\item{plotcov}{plot the GP covariances (default)}

\item{plotnoise}{plot the observational noise (default)}

\item{samples}{plot N samples from the GP}

\item{sigma}{variance level to plot}

\item{title}{plot title}

\item{legend}{plot legend}

\item{plotgradient}{use gradient graphics}

\item{...}{...}
}
\description{
Plots a gaussian process. Several boolean parameters for modifying the
 plot. By default plots the data, posterior mean and 95\% interval.
}
\keyword{internal}

