% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.gppack}
\alias{print.gppack}
\title{prints the two-sample GP summary}
\usage{
\method{print}{gppack}(x, ...)
}
\arguments{
\item{x}{the estimated gppack-object containing ctrlmodel, casemodel and nullmodel}

\item{...}{for compability}
}
\description{
prints the two-sample GP summary
}

