% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.gpsimple}
\alias{print.gpsimple}
\title{prints the gpsimple object}
\usage{
\method{print}{gpsimple}(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{for compatibility}
}
\description{
prints the gpsimple object
}
\keyword{internal}

