% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simulategp}
\alias{simulategp}
\title{Generate simulated GP models}
\usage{
simulategp(N = 100, reps = 3, obs = c(0, 5, 10, 15, 20), xs = seq(0, 20,
  0.2), filename = NULL, l.noise = 12, l.shape = 4, l.scale = 4,
  sigmaf.noise = 0.25, sigmaf.shape = 1.5, sigmaf.scale = 0.3)
}
\arguments{
\item{N}{number of gp's}

\item{reps}{replicate observations}

\item{obs}{observation timepoints}

\item{xs}{target timepoints}

\item{filename}{file to save the results}

\item{l.noise}{Noise model lengthscale}

\item{l.shape}{shape of lengthscale Gamma distribution}

\item{l.scale}{scale of lengthscale Gamma distribution}

\item{sigmaf.noise}{Noise model sigma.f}

\item{sigmaf.shape}{shape of sigma.f Gamma distribution}

\item{sigmaf.scale}{scale of sigma.f Gamma distribution}
}
\value{
List with
 \item{simdata}{Simulated datamatrix}
 \item{gps}{Simulated GPs}
}
\description{
Generate simulated GP models
}
\seealso{
\code{\link{simulategp.perturbed}}
}

