% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.gp}
\alias{summary.gp}
\title{prints the one-sample GP summary}
\usage{
\method{summary}{gp}(object, ...)
}
\arguments{
\item{object}{the estimated GP-object}

\item{...}{for compatibility}
}
\description{
identical to \code{\link{print.gp}}
}

