% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.gppack}
\alias{summary.gppack}
\title{prints the two-sample GP summary}
\usage{
\method{summary}{gppack}(object, ...)
}
\arguments{
\item{object}{the estimated gppack-object containing ctrlmodel, casemodel and nullmodel}

\item{...}{for compability}
}
\description{
identical to \code{\link{print.gppack}}
}

