% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.gpsimple}
\alias{summary.gpsimple}
\title{printss the gpsimple object}
\usage{
\method{summary}{gpsimple}(object, ...)
}
\arguments{
\item{object}{the object}

\item{...}{for compatibility}
}
\description{
identical to \code{\link{print.gpsimple}}
}
\keyword{internal}

