% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{toydata}
\alias{toydata}
\title{Toy time-series data for testing, contains two time-series, case and control}
\format{A list of two dataframes containing 24 (x,y)
 pairs both, i.e. toydata$ctrl and toydata$case}
\source{
randomly generated data
}
\usage{
data(toydata)
}
\description{
Toy time-series data for testing, contains two time-series, case and control
}
\keyword{datasets}

