% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{toygps}
\alias{toygps}
\title{Toy time-series model for testing}
\format{A \code{gppack} object}
\source{
prelearned model with \code{\link{gpr2sample}}
}
\usage{
data(toygps)
}
\description{
contains the GP models for the 'toydata'
}
\keyword{datasets}

